cmake: add support for Clang compilers
Check GCC and Clang compiler versions and exit with an error
when they are too old to provide working `std::format()`.

Use libc++ with Clang because it gained `std::format()` support
earlier than libstdc++.

Do not set -Werror=cast-function-type for Clang compilers.
This warning is triggered by the OpenSSL headers and can not be fixed.